requirements.txt
streamlit>=1.35.0
supabase>=2.5.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
pandas>=2.2.0
plotly>=5.22.0
python-dotenv>=1.0.0
requests>=2.32.0
